python/requirements.txt
ffmpeg-python
requests
numpy
opencv-python
imutils
fastapi
uvicorn[standard]
pydantic
jinja2
python-multipart
mediapipe
aiofiles
sqlalchemy
psycopg2
